requirements.txt
pyyaml>=6.0
demisto-sdk>=1.38.21
